+++ lib/src/compile.ts
@@ -4,6 +4,7 @@
 
 import * as p from 'path';
 import {Observable} from 'rxjs';
+import * as supportsColor from 'supports-color';
 
 import * as proto from './vendor/embedded-protocol/embedded_sass_pb';
 import {AsyncEmbeddedCompiler} from './async-compiler';
@@ -80,6 +81,9 @@ function newCompileStringRequest(
     case 'css':
       input.setSyntax(proto.Syntax.CSS);
       break;
+
+    default:
+      throw new Error(`Unknown options.syntax: "${options?.syntax}"`);
   }
 
   if (options?.url) input.setUrl(options.url.toString());
@@ -96,6 +100,23 @@ function newCompileRequest(
 ): proto.InboundMessage.CompileRequest {
   const request = new proto.InboundMessage.CompileRequest();
   request.setSourceMap(!!options?.sourceMap);
+  request.setAlertColor(options?.alertColor ?? !!supportsColor.stdout);
+  request.setAlertAscii(!!options?.alertAscii);
+  request.setQuietDeps(!!options?.quietDeps);
+  request.setVerbose(!!options?.verbose);
+
+  switch (options?.style ?? 'expanded') {
+    case 'expanded':
+      request.setStyle(proto.OutputStyle.EXPANDED);
+      break;
+
+    case 'compressed':
+      request.setStyle(proto.OutputStyle.COMPRESSED);
+      break;
+
+    default:
+      throw new Error(`Unknown options.style: "${options?.style}"`);
+  }
 
   for (const path of options?.loadPaths ?? []) {
     const importer = new proto.InboundMessage.CompileRequest.Importer();
@@ -250,7 +271,7 @@ function handleCompileResponse(
     const success = response.getSuccess()!;
     const result: CompileResult = {
       css: success.getCss(),
-      loadedUrls: [], // TODO(nex3): Fill this out
+      loadedUrls: success.getLoadedUrlsList().map(url => new URL(url)),
     };
 
     const sourceMap = success.getSourceMap();